switch password and id in munge_password
I think this is the less effective way to salt a password, BUT...
I had a big bug in my code! All the code that was calling this
function got the order of the params mixed up. So, to save heartache,
I just reversed the order of the parameters, and reversed how the were
used in the function. It should be functionally the same as before,
but better documented.

darcs-hash:20080618211454-84dde-4772135cc92f2a51faa072e4470da3c628e1510c.gz