remove favoritedrss
I've removed the favoritedrss action, and removed the links from the
favorited page.

Most-favorited is not a good candidate for an RSS format. RSS works
well for series that work sequentially -- new things always come in at
the top.

This just isn't the case with "most favorited". The favorited notices
list on Identi.ca is sorted by how many times a message is favorited.

What happens if a notice used to be on that list, then falls off
because other messages get more faves, then gets back on because it
gets faved again? Or if a notice moves "up" on the list?

darcs-hash:20081207165900-84dde-23afe73c9c61cfb259ec5dc0cac5030643b31f2c.gz