Revert "Fixed direct messaging: AjaxWebChannel is now using Action's methods."
This reverts commit 0f2c43bd040437b3e40c706d7c3f4ba163e94a71.

Making Channel a subclass of Action for no other reason than to let
the AjaxWebChannel do some output is the really, really wrong way to
do this. A Channel is not an Action.

I'll change AjaxWebChannel so it takes an Action as a constructor
paramater and uses that Action for its output. We do this for most
Widget subclasses and it makes sense here, too.