Unify feeds definition in actions
I got a little sick of trying to keep the export data and <head> links
synched in actions, so I made a common method, getFeeds(), which gets
the feeds for both. It returns an array of Feed objects, which know
about what their mime type is, title, location, all that jazz.

I changed the FeedList class so it handles the new Feed objects
instead of the old array of data.

I changed all the actions that show feeds (I think...) so that they
now use getFeeds() for all their feed needs.