Remove extranneous bottom margins on label; fix Safari layout bug.
This margin causes a layout bug in Safari whereby the Inbox and Outbox
pages have a notice form textarea that's too wide, causing it to slide
beneath the "Send" button on those pages. Removing this margin seems to
be completely innocuous in my tests, which covered Safari, Firefox,
Opera, and Internet Explorer 7. Can less be more? :)