Separation of themes (i.e., layout, typography) and skins (i.e., background images, colours).
A theme can have multiple skins.

Majority of the changes in this commit are due to restructuring the
path/files for themes and skins.

Both theme and skin will be set to 'default' if not set in
config.php.

This commit also allows each instance of this software to create
its own theme without having to override any style from the default
distribution.

Added Cloudy theme.