Takes care of http://laconi.ca/trac/ticket/1372 and a minor part of http://laconi.ca/trac/ticket/1367
One of the bugs was to replace foaf:knows with the weaker
sioc:follows. This is done, but foaf:knows is also retained when the
subscription is bidirectional.

Primary topic of the file is an Agent rather than a Person.

rdfs:comment replaced by bio:olb which is a fairly popular RDF
property for one-line biographies

rdf:about attribute to the <OnlineAccount> elements in a manner
consistent with the RSS 1.0 feed.

Changed the OnlineAccount's homepage property to an accountProfilePage
property. Will be added to FOAF spec soon (already widely used).

Added a <weblog> property from the Agent to their profile page.

Thanks to Toby Inkster for the patch.
http://buzzword.org.uk/2009/laconica-0.7.3-foafsioc.patch