Resolve Group Aliases in showgroup.php
For Example, let's say 'alias' was an alias for the group 'group'.
Previously, if you went to '/group/group' it'd work, but '/group/alias' it'd say "No Such Group".
This was untrue.

Now it checks aliases when it can't find a group with a given name.
If it finds one it redirects you to the original group.