IE has some issue with notices that are sent with file attachments. It doesn't like the XHR response with XHTML DTD. New notices without the file attachment work fine.
The rendered content (the anchor for the file attachment link) doesn't
appear to be the issue.

To fix this problem, I removed the XHTML DTD line from newnotice's XHR
response. This is unnecessary for text/xml outputs that's intended
for XHR responses any way. It just happens to fix an IE issue.

Still a mystery to me as to why it is particular to notices with file
attachments.