Fix for bug 1974: drop unnecessary define_syslog_variables() call from common.php
The call was moved to this spot in commit 3ea1119e500c23ee918569e2a58cc5abaa1d8a5a (previously init'd later in another func) but doesn't seem to be needed anyway. None of our code uses the variables that this function defines, just the constants -- which are already predefined without the call.

The function is deprecated in PHP 5.3 and gone in 6, so we may as well toss it now.