Fix for ticket #2286: [mobile] Text extends beyond border of repeat confirmation dialog floater box on iPhone
http://status.net/open-source/issues/2286

This bit of CSS was constricting the vertical size of the popup form for repeats:

.notice-options form {
 width:16px;
 height:16px;
}

I can only assume this was originally meant to constrain the mini inline AJAX forms to the size of the clickable buttons, but it doesn't make a difference to how those are displayed on iPhone, Android, or Opera Mini.
Removing the statement lets the popup form go to its natural size, covering the button.