Fix for ticket #2837: white screen when hitting 'make admin' button on group members list.
RedirectingAction->returnToArgs() has been renamed to returnToPrevious() to avoid confusion with the existing Action->returnToArgs which gives the arguments that would be used to pass to one of those. :)
Switching to the correct function call gets it working.

Thanks to Siebrand for catching it when I could debug it live!